abstract auth check
